move tests to test_internals/
